Improve admin panel for featured items and dashboards. Fix tests
