+++ lib/sanbase/dashboards/dashboard/dashboard.ex
@@ -118,6 +118,14 @@ defmodule Sanbase.Dashboards.Dashboard do
   @preload [:queries, :user, :featured_item, [queries: :user]]
   def default_preload(), do: @preload
 
+  @doc false
+  def changeset(%__MODULE__{} = dashboard, attrs) do
+    # Used in admin panel
+    dashboard
+    |> cast(attrs, @update_fields)
+    |> validate_required([:name])
+  end
+
   def create_changeset(%__MODULE__{} = dashboard, attrs) do
     dashboard
     |> cast(attrs, @create_fields)
